No more warnings

+++ apply.c
@@ -3052,7 +3052,7 @@ static int apply_one_fragment(struct apply_state *state,
 		     trailing != frag->trailing) && state->apply_verbosity > verbosity_silent)
 			fprintf_ln(stderr, _("Context reduced to (%ld/%ld)"
 					     " to apply fragment at %d"),
-				   leading, trailing, applied_pos+1);
+				   (long)leading, (long)trailing, applied_pos+1);
 		update_image(state, img, applied_pos, &preimage, &postimage);
 	} else {
 		if (state->apply_verbosity > verbosity_normal)
@@ -3208,7 +3208,7 @@ static int apply_fragments(struct apply_state *state, struct image *img, struct 
 	while (frag) {
 		nth++;
 		if (apply_one_fragment(state, img, frag, inaccurate_eof, ws_rule, nth)) {
-			error(_("patch failed: %s:%ld"), name, frag->oldpos);
+		  error(_("patch failed: %s:%ld"), name, (long)frag->oldpos);
 			if (!state->apply_with_reject)
 				return -1;
 			frag->rejected = 1;
